fairly error free thus far. If you exit, we kill the process if running. We stop when running, warn when not, start when we don't already have an xmlrpc server running and warn you when we do.

+++ lib/wXf/wXfxmlrpc/xmlrpc_server.rb
@@ -18,10 +18,11 @@ module WXf
     end
     
     def stop_server
+      return unless self.pid != nil
       Process.kill("KILL", self.pid)
+      @control.prnt_gen("Stopping xmlrpc server with process id: #{self.pid}")
       self.pid = nil
       self.shutdown
-      @control.prnt_gen("Stopping xmlrpc server with process id: #{self.pid}")
     end
     
     def mount(path, handler)
